rename api.rb to client.rb
