+++ lib/nylas/client.rb
